chore: moved accounts event watcher to accountsevent package
